[Feature #163042447] refactoring of code
