+++ package.json
@@ -3,34 +3,44 @@
   "version": "1.0.0",
   "description": "Crowd-source questions for a meetup",
   "main": "server.js",
-  "dependencies": {
-    "babel-core": "^6.26.3",
-    "babel-preset-es2015": "^6.24.1",
-    "body-parser": "^1.18.3",
-    "compression": "^1.7.3",
-    "express": "^4.16.4",
-    "helmet": "^3.15.0"
-  },
   "devDependencies": {
+    "@babel/cli": "^7.2.3",
+    "@babel/core": "^7.2.2",
+    "@babel/node": "^7.2.2",
+    "@babel/preset-env": "^7.2.3",
+    "@babel/register": "^7.0.0",
     "babel-cli": "^6.26.0",
-    "chai": "^4.2.0",
-    "chai-http": "^4.2.1",
+    "babel-eslint": "^10.0.1",
+    "babel-polyfill": "^6.26.0",
+    "babel-preset-env": "^1.7.0",
+    "babel-preset-es2015": "^6.24.1",
     "codeclimate-test-reporter": "^0.5.1",
     "coveralls": "^3.0.2",
     "eslint": "^5.11.0",
-    "eslint-config-airbnb": "^17.1.0",
     "eslint-config-airbnb-base": "^13.1.0",
     "eslint-plugin-import": "^2.14.0",
-    "eslint-plugin-jsx-a11y": "^6.1.2",
     "istanbul": "^0.4.5",
     "mocha": "^5.2.0",
     "mocha-lcov-reporter": "^1.3.0",
-    "nyc": "^13.1.0",
+    "nodemon": "^1.18.9",
+    "nyc": "^13.1.0"
+  },
+  "dependencies": {
+    "airbnb-browser-shims": "^3.1.0",
+    "babel-core": "^6.26.3",
+    "babel-istanbul": "^0.12.2",
+    "babel-register": "^6.26.0",
+    "body-parser": "^1.18.3",
+    "chai": "^4.2.0",
+    "chai-http": "^4.2.0",
+    "compression": "^1.7.3",
+    "express": "^4.16.4",
+    "helmet": "^3.15.0",
     "request": "^2.88.0",
     "supertest": "^3.3.0"
   },
   "scripts": {
-    "start": "node server.js",
+    "start": "node start.js",
     "test": "nyc mocha --exit --compilers js:babel-core/register ./test/**/*.js",
     "lint": "eslint --ignore-path .gitignore .",
     "coverage": "nyc --reporter=lcov --reporter=text-lcov --require babel-core/register  _mocha --exit test/*.js",
