+++ server/models/question.js
@@ -19,4 +19,4 @@ const questions = [
   },
 ];
 
-module.exports = questions;
+export default questions;
