+++ start.js
@@ -1,0 +1,5 @@
+require('babel-register')({
+  presets: ['env'],
+});
+
+module.exports = require('./server.js');
